+++ kernel/data-pipeline/core/src/test/java/org/apache/shardingsphere/data/pipeline/api/job/progress/yaml/YamlInventoryIncrementalJobItemProgressSwapperTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.it.data.pipeline.core.job.progress.yaml;
+package org.apache.shardingsphere.data.pipeline.api.job.progress.yaml;
 
 import org.apache.shardingsphere.data.pipeline.api.job.progress.InventoryIncrementalJobItemProgress;
 import org.apache.shardingsphere.data.pipeline.core.job.progress.yaml.YamlInventoryIncrementalJobItemProgress;
