+++ kernel/data-pipeline/core/src/test/resources/META-INF/services/org.apache.shardingsphere.data.pipeline.spi.ingest.position.PositionInitializer
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.test.it.data.pipeline.core.fixture.FixturePositionInitializer
+org.apache.shardingsphere.data.pipeline.core.ingest.position.FixturePositionInitializer
