+++ kernel/data-pipeline/core/src/test/resources/job-progress-all-finished.yaml
