+++ kernel/data-pipeline/core/src/test/resources/job-progress-running.yaml
